make compile independent of 'inputs_like'
'inputs_like' was a unnecessary argument to make the implementation
easier. But that led to a confusing interface. Now the compiled operator
is smart enough to do the inference of the input data types on its own.
It maintains a mapping from the input dtypes it saw to the executable.